refactor(projects): split Projects domain into Commands, Queries, RequestModels, ViewModels, Errors
Moves flat Projects/*.cs files into subfolders matching the documented domain layout (issue #192), and relocates the Projects-only SetCollectionRequest out of the legacy Controllers/RequestModels folder into Projects/RequestModels.
