+++ src/SheetMusic.Api/Projects/Commands/DisconnectSetFromProject.cs
@@ -6,7 +6,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SheetMusic.Api.Projects;
+namespace SheetMusic.Api.Projects.Commands;
 
 public class DisconnectSetFromProject(Guid projectId, Guid setId) : IRequest
 {
