+++ src/SheetMusic.Api/Projects/ProjectsController.cs
@@ -2,10 +2,13 @@ using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using SheetMusic.Api.Authorization;
-using SheetMusic.Api.Controllers.RequestModels;
 using SheetMusic.Api.Controllers.ViewModels;
 using SheetMusic.Api.CQRS.Query;
 using SheetMusic.Api.OData.MVC;
+using SheetMusic.Api.Projects.Commands;
+using SheetMusic.Api.Projects.Queries;
+using SheetMusic.Api.Projects.RequestModels;
+using SheetMusic.Api.Projects.ViewModels;
 using System.Linq;
 using System.Threading.Tasks;
