+++ src/SheetMusic.Api/Projects/Queries/GetProject.cs
@@ -7,7 +7,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SheetMusic.Api.Projects;
+namespace SheetMusic.Api.Projects.Queries;
 
 public class GetProject(string identifier) : IRequest<Project>
 {
