+++ src/SheetMusic.Api/Projects/Queries/GetSetsForProject.cs
@@ -8,7 +8,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace SheetMusic.Api.Projects;
+namespace SheetMusic.Api.Projects.Queries;
 
 public class GetSetsForProject(Guid projectId) : IRequest<List<SheetMusicSet>>
 {
