+++ src/SheetMusic.Api/Projects/RequestModels/NewProjectRequest.cs
@@ -1,14 +1,14 @@
 using System;
 
-namespace SheetMusic.Api.Projects;
+namespace SheetMusic.Api.Projects.RequestModels;
 
-public class UpdateProjectRequest
+public class NewProjectRequest
 {
     public string Name { get; set; } = null!;
-
-    public string? Comments { get; set; }
 
     public DateTime StartDate { get; set; }
 
     public DateTime EndDate { get; set; }
+
+    public string? Comments { get; set; }
 }
